Cyclical Coordinate Descent

Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.

RiskBudgeting.ccd — Function

ccd(cov, b, [max_iter], [tol], [bounds])

cov::AbstractMatrix Covariance matrix
b::AbstractVector{Float} Risk budgeting vector
max_iter::Int Number of iterations for cyclical coordinate descent
tol::Float The minimum tolerance of the result
bounds::Bool Whether to run bounds checks or not

Calculates the solution of the risk budgeting portfolio with Cyclical Coordinate Descent given the covariance matrix and the risk partitions between the assets.

External links

Griveau-Billion, Théophile and Richard, Jean-Charles and Roncalli, Thierry, A Fast Algorithm for Computing High-Dimensional Risk Parity Portfolios (September 1, 2013). doi: 10.2139/ssrn.2325255

RiskBudgeting.fastccd — Function

fastccd(cov, b, [max_iter], [tol], [bounds])

cov::AbstractMatrix Covariance matrix
b::AbstractVector{Float} Risk budgeting vector
max_iter::Int Number of iterations for cyclical coordinate descent
tol::Float The minimum tolerance of the result
bounds::Bool Whether to run bounds checks or not

A faster solution finder of the risk budgeting portfolio with Cyclical Coordinate Descent given the covariance matrix and the risk partitions between the assets.

External links

Choi, J., & Chen, R. (2022). Improved iterative methods for solving risk parity portfolio, Journal of Derivatives and Quantitative Studies, 30(2) doi: 10.48550/arXiv.2203.00148