Solvers

Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.

Cyclical Coordinate Descent

RiskBudgeting.ccd — Function

ccd(cov, b, [max_iter], [tol], [bounds])

cov::AbstractMatrix Covariance matrix
b::AbstractVector{Float} Risk budgeting vector
max_iter::Int Number of iterations for cyclical coordinate descent
tol::Float The minimum tolerance of the result
bounds::Bool Whether to run bounds checks or not

Calculates the solution of the risk budgeting portfolio with Cyclical Coordinate Descent given the covariance matrix and the risk partitions between the assets.

External links

Griveau-Billion, Théophile and Richard, Jean-Charles and Roncalli, Thierry, A Fast Algorithm for Computing High-Dimensional Risk Parity Portfolios (September 1, 2013). doi: 10.2139/ssrn.2325255

RiskBudgeting.fastccd — Function

fastccd(cov, b, [max_iter], [tol], [bounds])

cov::AbstractMatrix Covariance matrix
b::AbstractVector{Float} Risk budgeting vector
max_iter::Int Number of iterations for cyclical coordinate descent
tol::Float The minimum tolerance of the result
bounds::Bool Whether to run bounds checks or not

A faster solution finder of the risk budgeting portfolio with Cyclical Coordinate Descent given the covariance matrix and the risk partitions between the assets.

External links

Choi, J., & Chen, R. (2022). Improved iterative methods for solving risk parity portfolio, Journal of Derivatives and Quantitative Studies, 30(2) doi: 10.48550/arXiv.2203.00148

Newton's Method

RiskBudgeting.newton — Function

newton(cov, b, [max_iter], [tol], [bounds])

cov::AbstractMatrix Covariance matrix
b::AbstractVector{Float} Risk budgeting vector
max_iter::Int Number of iterations for cyclical coordinate descent
tol::Float The minimum tolerance of the result
bounds::Bool Whether to run bounds checks or not

Solution finder of the risk budgeting portfolio with Newton's method given the covariance matrix and the risk partitions between the assets (also known as Spinu's algorithm).

External links

Spinu, Florin, An Algorithm for Computing Risk Parity Weights (July 30, 2013). doi: 10.2139/ssrn.2297383

Note

Note that the modified Spinu's algorithm (fastnewton) proposed by Choi, J., & Chen, R. can obtain negative weights, which according to the constraints of risk budgeting should not occur. In addition, the identical solver the authors use, the hybrj routine, does not have a stable version in Julia at this moment.

RiskBudgeting.fastnewton — Function

fastnewton(cov, b, [tol], [bounds])

cov::AbstractMatrix Covariance matrix
b::AbstractVector{Float} Risk budgeting vector
tol::Float The minimum tolerance of the result
bounds::Bool Whether to run bounds checks or not

A faster solution finder of the risk budgeting portfolio based on Newton's method given the covariance matrix and the risk partitions between the assets. The core concept is that the solution will converge faster if the initial guess of the weights is closer to the solution in comparision to Spinu's algorithm.

External links

Choi, J., & Chen, R. (2022). Improved iterative methods for solving risk parity portfolio, Journal of Derivatives and Quantitative Studies, 30(2) doi: 10.48550/arXiv.2203.00148